Group tests into pure multipart & form data
